+++ shifter/engine/provisioner/tests/test_edge_cases.py
@@ -116,8 +116,6 @@ class TestConfigBoundaryValues:
             vpc_id="vpc-12345",
             vpc_cidr="10.1.0.0/16",
             route_table_id="rtb-12345",
-            kali_security_group_id="sg-kali",
-            victim_security_group_id="sg-victim",
             instance_profile_name="profile",
             kali_ami_id="ami-kali",
             victim_ami_id="ami-ubuntu",